Update 23/08/2021 - Light sources and damge, player controller, beginnings of animation
I created a player controller, with health and the ability to be damaged by light sources. With this we already have an MVP for some kind of stealth game where you need to stay out of the way of lights, but our game obviously needs more than that. The most important features I need to make are the lives system, game over screen, and the ability to disable lights. These should be relatively easy but may take time. I also want to add some other mechanics to the puzzles, such as shifting/physics-based light sources and movable objects to block light.