Update 23/08/2021: Interaction, respawning, pausing, global level data
Created the following:
Player MonoBehaviour for easily linking different scripts needed for the player’s systems to work, and for easily referencing the player
Added code to the player animation controller, to properly represent the direction the player is moving in, and animate their movement speed appropriately
GameStateHandler for managing pausing and resuming
Interact function for disabling lights
Fully implemented respawn function
LevelData script for managing variables and win state for the current level, with a handy singleton that should update appropriately when a new level is loaded
SpawnPoint script for respawning the player, with a handy singleton that should update appropriately when a new level is loaded

Started creating the following:
Functions for menu buttons, such as restarting the level or quitting the game
Pause menu
WinZone that the player must reach in order to complete the level.

Notable problems to fix
Player does not have the movement strength to move up 45* slopes. I attempted implementing a system that obtains the normal of whatever the player is standing on, and adjusts their movement input to move relative to its angle, so the movement isn’t fighting the slope by going directly forward. I was not able to complete it, however, due to a variety of aggravating problems and other obligations I had to deal with.
Player does not register they are grounded when standing on certain slopes. I tried fixing this by creating a SphereCast that covers the whole area underneath the player collider, but this caused all kinds of issues where the player controller would never register it was grounded, for reasons I was unable to understand considering I have created very similar functions that have worked fine.
Lights do not change appearance when turned on and off. This is currently confusing from a gameplay perspective, but I could not figure out an elegant method to fix this.

For the most part, despite infuriating bugs and annoyances, I am pleased with the work I've done. I have pretty much all the important elements for the MVP. That being said, there is plenty more work to do.